\name{X}
\alias{X}
\docType{data}
\title{
A simulated data X.
}
\description{
For exmaple run.
}
\usage{data("X")}
\format{
  The format is:
 int [1:200, 1:100] 1 1 1 0 0 0 0 0 0 2 ...
 - attr(*, "dimnames")=List of 2
  ..$ : NULL
  ..$ : chr [1:100] "V1" "V2" "V3" "V4" ...
}

\examples{
data(X)
}
\keyword{datasets}
